+++ src/NServiceBus.Transport.Msmq.AcceptanceTests/When_deferring_a_message_that_cant_be_stored.cs
@@ -50,7 +50,7 @@
                 {
                     endpointConfiguration.SendFailedMessagesTo(Conventions.EndpointNamingConvention(typeof(ErrorSpy)));
                     var transport = endpointConfiguration.ConfigureTransport<MsmqTransport>();
-                    transport.DelayedDelivery = new DelayedDeliverySettings(new FaultyTimeoutStorage(transport.DelayedDelivery.TimeoutStorage), 2);
+                    transport.DelayedDelivery = new DelayedDeliverySettings(new FaultyDelayedMessageStore(transport.DelayedDelivery.DelayedMessageStore), 2);
                 });
             }
 
@@ -99,16 +99,16 @@
         {
         }
 
-        class FaultyTimeoutStorage : ITimeoutStorage
+        class FaultyDelayedMessageStore : IDelayedMessageStore
         {
-            ITimeoutStorage impl;
+            IDelayedMessageStore impl;
 
-            public FaultyTimeoutStorage(ITimeoutStorage impl)
+            public FaultyDelayedMessageStore(IDelayedMessageStore impl)
             {
                 this.impl = impl;
             }
 
-            public Task Initialize(string endpointName, CancellationToken cancellationToken) => impl.Initialize(endpointName, cancellationToken);
+            public Task Initialize(string endpointName, TransportTransactionMode transactionMode, CancellationToken cancellationToken) => impl.Initialize(endpointName, transactionMode, cancellationToken);
 
             public Task<DateTimeOffset?> Next() => impl.Next();
 
@@ -119,7 +119,7 @@
 
             public Task<bool> Remove(TimeoutItem entity) => impl.Remove(entity);
 
-            public Task<bool> BumpFailureCount(TimeoutItem timeout) => impl.BumpFailureCount(timeout);
+            public Task<bool> IncrementFailureCount(TimeoutItem timeout) => impl.IncrementFailureCount(timeout);
 
             public Task<TimeoutItem> FetchNextDueTimeout(DateTimeOffset at) => impl.FetchNextDueTimeout(at);
         }
